backend/CineBrain/chatBoot/serializers.py
from rest_framework import serializers
from .models import Room, Message


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = '__all__'    
    